requirements.txt
asgiref
charset-normalizer
coreapi
coreschema
Django==3.1.5
django-baton==2.3.0
django-ckeditor==6.3.2
django-cors-headers==3.10.1
django-filter==21.1
django-js-asset==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt
drf-yasg
MarkupSafe
packaging
Pillow
psycopg2
# psycopg2-binary==2.7.4
PyJWT==v1.7.1
requests
python-decouple
Unidecode
xmltodict
django-debug-toolbar
django-rest-passwordreset
uWSGI
pytz
django-baton
celery
django-celery-results
django-celery-beat
redis
gunicorn
django-imagekit
easy-thumbnails
mongoengine
pymongo
djongo
pylint-mongoengine
